# examples/acme-bank/gateway/src/main/resources/openapi/bank-api.yaml
openapi: 3.0.3
info:
  title: acme-bank API
  version: 1.0.0
  description: Money-movement API. Spec-first — this contract is the source of truth.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /v1/transfers:
    post:
      tags: [transfers]
      operationId: createTransfer
      summary: Initiate a transfer (idempotent via Idempotency-Key)
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTransferRequest' }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferAccepted' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { description: Unauthorized }
        '409': { $ref: '#/components/responses/Problem' }
    get:
      tags: [transfers]
      operationId: listTransfers
      parameters:
        - { in: query, name: accountId, schema: { type: string } }
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 0 } }
        - { in: query, name: size, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferPage' }
  /v1/transfers/{id}:
    get:
      tags: [transfers]
      operationId: getTransfer
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferView' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/accounts:
    post:
      tags: [accounts]
      operationId: openAccount
      summary: Open an account (idempotent via Idempotency-Key), optionally with an opening deposit
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OpenAccountRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountView' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { description: Unauthorized }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/accounts/{id}:
    get:
      tags: [accounts]
      operationId: getAccount
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountView' }
        '404': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/accounts/{id}/balance:
    get:
      tags: [accounts]
      operationId: getAccountBalance
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Money' }
        '404': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/accounts/{id}/statement:
    get:
      tags: [accounts]
      operationId: getAccountStatement
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 0 } }
        - { in: query, name: size, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatementPage' }
        '404': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Problem:
      description: RFC 9457 problem+json
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Money:
      type: object
      required: [value, asset]
      properties:
        value: { type: string, example: "100.00", description: "Exact decimal string, no float" }
        asset: { type: string, example: USD }
    CreateTransferRequest:
      type: object
      required: [sourceAccountId, destinationAccountId, amount]
      properties:
        sourceAccountId: { type: string }
        destinationAccountId: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        reference: { type: string }
    TransferAccepted:
      type: object
      required: [transferId, status]
      properties:
        transferId: { type: string }
        status: { type: string }
    TransferView:
      type: object
      required: [transferId, status, amount, sourceAccountId, destinationAccountId]
      properties:
        transferId: { type: string }
        status: { type: string, enum: [REQUESTED, SCREENING, APPROVED, POSTING, COMPLETED, FAILED] }
        amount: { $ref: '#/components/schemas/Money' }
        sourceAccountId: { type: string }
        destinationAccountId: { type: string }
        failureReason: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TransferPage:
      type: object
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/TransferView' } }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
    OpenAccountRequest:
      type: object
      required: [ownerName, asset]
      properties:
        ownerName: { type: string }
        asset: { type: string, example: USD }
        initialDeposit: { $ref: '#/components/schemas/Money' }
    AccountView:
      type: object
      required: [accountId, iban, status]
      properties:
        accountId: { type: string }
        iban: { type: string }
        status: { type: string, enum: [OPEN, FROZEN, CLOSED] }
    StatementLine:
      type: object
      required: [postedAt, signedAmount, runningBalance]
      properties:
        postedAt: { type: string }
        counterpartyAccountId: { type: string }
        signedAmount: { type: string, description: "Exact decimal string, no float" }
        runningBalance: { type: string, description: "Exact decimal string, no float" }
    StatementPage:
      type: object
      properties:
        accountId: { type: string }
        page: { type: integer }
        size: { type: integer }
        lines: { type: array, items: { $ref: '#/components/schemas/StatementLine' } }
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }